server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.20
Pillow==10.4.0
numpy==1.26.4
paddleocr==2.8.1
manga-ocr==0.1.14
